requirements.txt
tweety-ns>=1.0.0
selenium>=4.15.2
requests==2.31.0
pyyaml==6.0.1
apscheduler==3.10.1
pytz==2023.3
python-dotenv==1.0.0
webdriver-manager>=4.0.1